Move Class Widget Catalog to widget_catalog.dart
